# Conflicts:
#	composeApp/build.gradle.kts
#	composeApp/src/commonMain/kotlin/cn/allin/App.kt
#	composeApp/src/commonMain/kotlin/cn/allin/ui/fileMamager/FileManagerViewModel.kt
#	composeApp/src/commonMain/kotlin/cn/allin/ui/fileMamager/RouteFileManager.kt
#	composeApp/src/commonMain/kotlin/cn/allin/ui/fileMamager/TransferManagerViewModel.kt

+++ admin/src/jsMain/kotlin/cn/allin/ui/AdminPageTable.kt
@@ -24,7 +24,7 @@ data class PageParams(
 
 external interface AdminTableProps : Props {
     var table: Table<*>
-    var pageCount: Int?
+    var pageCount: Long?
     var page: PageParams
     var onPage: (Int) -> Unit
     var onPageParams: (PageParams) -> Unit
